% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics.R
\name{aes_map}
\alias{aes_map}
\alias{aes_map.default}
\alias{aes_map<-}
\alias{aes_map<-.default}
\title{Set and retrieve mappings for facile objects.}
\usage{
aes_map(x, aes_name = NULL, ...)

\method{aes_map}{default}(x, aes_name = NULL, ...)

aes_map(x, aes_name = NULL) <- value

\method{aes_map}{default}(x, aes_name = NULL) <- value
}
\arguments{
\item{x}{an orbitrary object}

\item{aes_name}{optional aes_name you want to individually set or retrieve,
otherweise entire aesthetic map list is set/returned}
}
\value{
the getter returns the whole aesthetic map when \code{is.null(aes_name)},
or the specific mapping for the named aesthetic. The setter sets the
specific \code{aes_name} when specified, otherwise sets the aesthetics in
the named \code{value} list
}
\description{
\itemize{
\item \code{aes_map} retrieves all aesthetic maps, or the map for an individual
aesthetic when \code{aes_name} is specified.
\item \code{aes_map()<-} resets the map, wholesale.
}
}
\examples{
x <- data.frame(a = 1:5, b = letters[1:5])
aes_map(x, "color") <- create_color_map(x$b)
aes_map(x, "size") <- "all sizes"
aes_map(x, "color")

# simultaneously updates color map and adds shape map
aes_map(x) <- list(color = create_color_map(x$b, "Set3"),
                   shape = "all shapes")
aes_map(x)
}
\seealso{
\code{\link[=create_color_map]{create_color_map()}}
}
