% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-rankings.R
\name{as.GeneSetDb}
\alias{as.GeneSetDb}
\title{Convert a feature descriptor into a GeneSetDb}
\usage{
as.GeneSetDb(x, ...)
}
\description{
This is mainly for feature rankings, but could be something else?
}
\examples{
pca.gdb <- FacileData::exampleFacileDataSet() \%>\%
  filter_samples(indication == "CRC") \%>\%
  fpca() \%>\%
  ranks() \%>\%
  as.GeneSetDb()
}
