% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdge-advanced.R
\name{as_fdge}
\alias{as_fdge}
\title{FacileAnalysis wrapper for a manually performed differential expression.'}
\usage{
as_fdge(x, ...)
}
\description{
This allows an analyst to perform a differential expression analysis
outside of the facileverse, and wrap it into an object that can be used
with the viz,report,shine functions.
}
