% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocbox.R, R/fdge.R
\name{biocbox}
\alias{biocbox}
\alias{biocbox.FacileDGEModelDefinition}
\alias{biocbox.FacileDGEResult}
\title{Materializes the bioc assay container to use to run a dge test.}
\usage{
biocbox(x, ...)

\method{biocbox}{FacileDGEModelDefinition}(x, assay_name, method = NULL,
  dge_methods = NULL, filter = "default",
  with_sample_weights = FALSE, prior_count = 3, ...)

\method{biocbox}{FacileDGEResult}(x, ...)
}
\arguments{
\item{x}{The FacileAnalysisResult}

\item{...}{passed down to modeling functions where appropriate.}

\item{assay_name}{the name of the assay to pull data for}

\item{method}{the name of the dge method that will be used. This will dictate
the post-processing of the data}

\item{filter}{A filtering policy to remove unintereesting genes.
If \code{"default"} (which is the default), then \code{\link[edgeR:filterByExpr]{edgeR::filterByExpr()}} is
used if we are materializing a \code{DGEList}, otherwise lowly expressed
features are removed in a similarly "naive" manner. This can,
alternatively, be a character vector that holds the names of the features
that should be kept. Default value: \code{"default"}.}

\item{with_sample_weights}{Some methods that leverage the limma pipeline,
like \code{"voom"}, \code{"limma"}, and \code{"limma-trend"} can leverage sample (array)
quality weights to downweight outlier samples. In the case of
\code{method == "voom"}, we use \code{\link[limma:voomWithQualityWeights]{limma::voomWithQualityWeights()}}, while the
rest use \code{\link[limma:arrayWeights]{limma::arrayWeights()}}. The choice of \code{method} determines which
sample weighting function to sue. Defaults to \code{FALSE}.}

\item{prior_count}{The pseudo-count to add to count data. Used primarily
when running the \code{limma-trend} method on count (RNA-seq) data.}

\item{sample_info}{a \code{facile_frame} that enumerates the samples to fetch
data for, as well as the covariates used in downstream analysis}
}
\value{
a DGEList or EList with assay data in the correct place, and all of
the covariates in the \code{$samples} or \code{$targerts} data.frame that are requied
to test the model in \code{mdef}.
}
\description{
Bioconductor assay containers need to be materialized for certain analyses,
like differential gene expression. They also may need to be materialized
from a result of something generated within some FacileAnalysis.
}
\section{Linear Model Definitions}{

This function accepts a model defined using using \code{\link[=fdge_model_def]{fdge_model_def()}} and
creates the appropriate Bioconductor assay container to test the model
given the \code{assay_name} and dge \code{method} specified by the user.

This function only supports RNA-seq like data, ie. data from assays from
these \code{assay_type}s:
\itemize{
\item \code{rnaseq}: assumed to be "vanilla" bulk rnaseq gene counts
\item \code{umi}: data from bulk rnaseq, UMI data, like quantseq
\item \code{tpm}: TPM values. These will be \code{log2(TPM + prior_count)} transformed,
then differentially tested using the limma-trended pipeline
}

TODO: support affymrna, affymirna, etc. assay types
}

\section{FacileDGEResult}{

Given a FacileDGEResult, we can re-materialize the Bioconductor assay
container used within the differential testing pipeline used from \code{\link[=fdge]{fdge()}}.
}

