% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{compare}
\alias{compare}
\title{Compares two (or more(?)) FacileAnalysisResults against each other.}
\usage{
compare(x, y, ...)
}
\description{
This was initially motivated by the desire to compare the results of two
differential expressio analyses against each other. Extending the idea
to comparing GSEA results (using Thomas' enrichmentmap idea) is a natural
extension. There may be othres.
}
\details{
Not every FacileAnalysisResult can be compared, I guess, in which case it
we will throw an error
}
