% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{contrast}
\alias{contrast}
\title{Extract the contrast defined.}
\usage{
contrast(x, ...)
}
\arguments{
\item{x}{a facileanalysis result / FacileLinearModelDefinition}
}
\description{
Extract the contrast defined.
}
