% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_aes_map.R
\name{create_color_map}
\alias{create_color_map}
\title{Maps colors to categorical values}
\usage{
create_color_map(vals, map = NULL)
}
\description{
Maps colors to categorical values
}
\section{Categorical Map}{

Map can be a RColorBrewer name, or a vector of colors. Colors will be
recycled if \code{length(map) <}
If \code{vals}, map can be is categoricasl
}

