% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{design}
\alias{design}
\title{Extract the design matrix from objects used in a linear modeling step}
\usage{
design(x, ...)
}
\description{
Extract the design matrix from objects used in a linear modeling step
}
