% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdge-interact.R
\name{.data_dge_feature}
\alias{.data_dge_feature}
\title{Extracts expression data for a feature across the samples use for the test.}
\usage{
.data_dge_feature(x, feature_id, batch_correct = TRUE, prior.count = 0.1, ...)
}
\description{
Extracts expression data for a feature across the samples use for the test.
}
