% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdge-compare.R
\name{.interaction_fdge}
\alias{.interaction_fdge}
\title{Helper function to run an interaction model to generate statistics when
comparing two Ttest models.}
\usage{
.interaction_fdge(x, y, ...)
}
\description{
This function is only to be called from within compare.FacileTTestDGEResult.
}
\details{
If we can't generate an interaction result, this will return NULL.
}
