% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{mutate.iresult}
\alias{mutate.iresult}
\alias{select.iresult}
\alias{filter.iresult}
\alias{summarise.iresult}
\alias{summarize.iresult}
\alias{arrange.iresult}
\alias{group_by.iresult}
\title{ImmersiveAnalysis result objects can be maniuplated with dplyr verbs}
\usage{
\method{mutate}{iresult}(.data, ...)

\method{select}{iresult}(.data, ...)

\method{filter}{iresult}(.data, ...)

\method{summarise}{iresult}(.data, ...)

\method{summarize}{iresult}(.data, ...)

\method{arrange}{iresult}(.data, ...)

\method{group_by}{iresult}(.data, ...)
}
\description{
ImmersiveAnalysis result objects can be maniuplated with dplyr verbs
}
