% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{faro}
\alias{faro}
\title{Extract the internal (F)acile(A)nalysis(R)esult (O)bject from a container.}
\usage{
faro(x, ...)
}
\value{
an innert FacileAnalysisResult
}
\description{
\code{FacileAnalysisResult} object can be stored within a
\code{ReactiveFacileAnalysis}, or a \code{FacileMultiAnalysisResult} (which is a
collection of FacileAnalysisResults returned from a "multi-analysis" view,
like \code{\link[=fDgeSeaAnalysis]{fDgeSeaAnalysis()}})
}
\details{
Due to the potential hybrid-nature of analysis workflow, we are not always
certain that a FacileAnalysisResult used by a module that did not create it
(ie. the \code{dgeres} object in the \code{\link[=fdgeView]{fdgeView()}} is a
\code{ReactiveFacileAnalysisResult}, or simply an innert \code{FacileAnalysisResult}.

So that we don't have to always check for these things, we call this method
so that we get access to the \code{FacileAnalysisResult} object itself.

As an aside, I know this is a bit of a cop-out to doing full-blown OO. I
feel like if I were to implement each of the specific \code{*FacileAnalysisResult}
methods for its sister \code{ReactiveFacile*AnalysisResult}, then this probably
wouldn't be necessary ... I think, anyway.
}
