% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdge-helpers.R
\name{is_ttest}
\alias{is_ttest}
\alias{is_anova}
\title{Checks to see if fdge analysis results are from a ttest or anova}
\usage{
is_ttest(x, ...)

is_anova(x, ...)
}
\description{
Checks to see if fdge analysis results are from a ttest or anova
}
