% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdge_biocbox.R
\name{fdge_biocbox}
\alias{fdge_biocbox}
\title{Materializes the bioc assay container to use to run a dge test.}
\usage{
fdge_biocbox(mdef, assay_name, method, dge_methods = NULL,
  filter = NULL, prior_count = 1, ...)
}
\arguments{
\item{assay_name}{the name of the assay to pull data for}

\item{method}{the name of the dge method that will be used. This will dictate
the post-processing of the data}

\item{sample_info}{a \code{facile_frame} that enumerates the samples to fetch
data for, as well as the covariates used in downstream analysis}
}
\value{
a DGEList or EList with assay data in the correct place, and all of
the covariates in the \code{$samples} or \code{$targerts} data.frame that are requied
to test the model in \code{mdef}.
}
\description{
This function accepts a model defined using using \code{\link[=fdge_model_def]{fdge_model_def()}} and
creates the appropriate Bioconductor assay container to test the model
given the \code{assay_name} and dge \code{method} specified by the user.
}
\details{
This function only supports RNA-seq like data, ie. data from assays from
these \code{assay_type}s:
\itemize{
\item \code{rnaseq}: assumed to be "vanilla" bulk rnaseq gene counts
\item \code{umi}: data from bulk rnaseq, UMI data, like quantseq
\item \code{tpm}: TPM values. These will be \code{log2(TPM + prior_count)} transformed,
then differentially tested using the limma-trended pipeline
TODO: support affymrna, affymirna, etc. assay types
}
}
