% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdge.R
\name{fdge_methods}
\alias{fdge_methods}
\title{A table of assay_type,dge_method combination parameters}
\usage{
fdge_methods(assay_type = NULL, on_missing = c("error", "warning"))
}
\arguments{
\item{assay_type}{An optional string specifying the valid dge methods for
a given assay type.}
}
\value{
A tibble of assay_type -> method and parameter associtiations. If
\code{assay_type}is not \code{NULL}, this will be filtered to the associations
valid only for that \code{assay_type}. If none are found, this will be a
0-row tibble.
}
\description{
This table is used to match the assay_type,dge_method combination with
the appropriate bioc container class \code{bioc_class}, and default edgeR/limma
model fitting params.
}
