% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdge_model_def-shiny.R
\name{fdgeModelDef}
\alias{fdgeModelDef}
\title{A shiny module that generates a linear model definition via \code{fdge_model_def}.}
\usage{
fdgeModelDef(input, output, session, rfds, ..., .reactive = TRUE)
}
\value{
A \code{FacileDGEModelDefinition} object, the output from
\code{\link[=fdge_model_def]{fdge_model_def()}}.
}
\description{
\preformatted{model_info <- fds \%>\%
  filter_samples(indication == "BLCA") \%>\%
  fdge_model_def(covariate = "sample_type",
                 numer = "tumor",
                 denom = "normal",
                 fixed = "sex")
}
}
