% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdge-shiny.R
\name{fdgeRun}
\alias{fdgeRun}
\title{Shiny module to configure and run fdge over a predefined model.}
\usage{
fdgeRun(input, output, session, rfds, model, with_gsea = FALSE, ...,
  debug = FALSE, .reactive = TRUE)
}
\arguments{
\item{model}{A linear model definition. Can be either a "naked"
\code{FacileDgeModelDefinition} that is returned from a call to
\code{\link[=fdge_model_def]{fdge_model_def()}}, or the \code{ShinyDGEModelDefinition} object returned
from the \code{\link[=fdgeModelDefRun]{fdgeModelDefRun()}} module.}

\item{with_gsea}{Include option to run a GSEA?}

\item{...}{passed into \code{\link[=fdge]{fdge()}}}
}
\value{
A list of stuff. \code{$result} holds the \code{FacileDgeAnalysisResult}
wrapped in a \code{reactive()}.
}
\description{
Shiny module to configure and run fdge over a predefined model.
}
