% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdge-shiny.R
\name{fdgeView}
\alias{fdgeView}
\title{Shiny module that interacts with result of fdgeRun module}
\usage{
fdgeView(input, output, session, rfds, result, ...,
  feature_selection = session$ns("volcano"),
  sample_selection = session$ns("samples"), debug = FALSE)
}
\arguments{
\item{result}{The result from calling \code{\link[=fdge]{fdge()}}.}
}
\description{
Provides an interactive view over the result of running the \code{\link[=fdgeRun]{fdgeRun()}}
module.
}
