% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdge-shiny.R
\name{fdgeViewResult}
\alias{fdgeViewResult}
\title{Shiny module that interacts with result of fdgeRun module}
\usage{
fdgeViewResult(input, output, session, rfds, fdge_result, ...)
}
\arguments{
\item{fdge_result}{The result from calling \code{\link[=fdge]{fdge()}}.}
}
\description{
Provides an interactive view over the result of running the \code{\link[=fdgeRun]{fdgeRun()}}
module.
}
