% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{feature_lookup}
\alias{feature_lookup}
\title{The primary key for a feature is ensembl id, uniprotid, etc, but somtimes you want to look up by gene names}
\usage{
feature_lookup(x, reference, assay_name = NULL, feature_name = "name", ...)
}
\arguments{
\item{x}{a character list of feature names/symbols}

\item{reference}{an object you are looking up in, like a FacileAnalysisResult, facile_frame, fds, ...}
}
\description{
The primary key for a feature is ensembl id, uniprotid, etc, but somtimes you want to look up by gene names
}
\examples{
feature_lookup(c("MCM2", "ENSG00000088930"), FacileData::an_fds())
}
