% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R, R/ffsea.R
\name{features}
\alias{features}
\alias{features.FacileFseaAnalysisResult}
\title{Returns a table of information about the features used in the analysis.}
\usage{
features(x, ...)

\method{features}{FacileFseaAnalysisResult}(x, ...)
}
\arguments{
\item{x}{A FacileAnalysisResult}
}
\value{
a tibble with containing feature_id, feature_type, and whatever other
columns that were used in the analysis.
}
\description{
The individual FacileAnalysis modules will define their own implementations
of this method. Think of "features" as being the things that the analysis
is generating statistics over.
}
\section{Feature Set Enrichment Analysis}{

What are the features of a feature-set enrichment analysis (\code{\link[=ffsea]{ffsea()}})?
Aren't they the gene sets, and not the individual genes themselves?
There is crappy support for this, for now.

It is a meta-something type of thing. The genesets are the features, but
they are also made up of their own features. We most often think of genesets
as consisting of genes, but perhaps we can imagine a feature set that
consists of motifs ... or sometihng.
}

