% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffsea-meta.R
\name{ffsea_methods}
\alias{ffsea_methods}
\title{Returns the types of featureset enrichment methods available over different
types of results.}
\usage{
ffsea_methods(x, ...)
}
\description{
Returns the types of featureset enrichment methods available over different
types of results.
}
