% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffsea-shiny.R
\name{ffseaGadget}
\alias{ffseaGadget}
\title{Full interactive executation of GSEA, soup to nuts.}
\usage{
ffseaGadget(x, title = "Feature Set Enrichment Analysis", height = 800,
  width = 1000, ...)
}
\arguments{
\item{x}{A FacileAnalysisResult that has an implemented \code{ffsea.*} method}
}
\description{
I think this has to be loaded with an already run FacileAnalysisResult
that we have an ffsea method for, ie. configure a GSEA from an already baked
DGE or PCA AnalysisResult.
}
\examples{

gdb <- multiGSEA::getMSigGeneSetDb("h", "human", id.type = "entrez")
dge.ttest <- FacileData::exampleFacileDataSet() \%>\%
  FacileData::filter_samples(indication == "CRC") \%>\%
  fdge_model_def(covariate = "sample_type",
                 numer = "tumor", denom = "normal", fixed = "sex") \%>\%
  fdge(method = "voom")
gsea.ttest <- ffsea(dge.ttest, gdb, methods = c("cameraPR", "fgsea"))
if (interactive()) {
  shine(ttest.gsea)
}
}
