% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgadget.R
\name{fgadget}
\alias{fgadget}
\title{Takes care of the tedium of wrapping a facilemodule to run as a gadget.}
\usage{
fgadget(x, samples = NULL, module, ui = NULL, ...)
}
\arguments{
\item{x}{A FacileDataStore}

\item{samples}{an optional subset of samples to launch the gadget over}

\item{module}{the to launch as a gadget}

\item{ui}{An optional UI function for the gadget. If none is provided,
it will default to calling \code{moduleUI()}

@examples
\dontrun{
 fds <- FacileData::exampleFacileDataSet()
 stuff <- fgadget(fds, fdgeAnalysis)
 }}
}
\description{
Takes care of the tedium of wrapping a facilemodule to run as a gadget.
}
