% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpca-shiny.R
\name{fpcaGadget}
\alias{fpcaGadget}
\title{Perform a PCA analysis on a (subset) of a FacileDataStore}
\usage{
fpcaGadget(x, title = "Principal Components Analysis", viewer = "browser", ...)
}
\description{
Interactively runs PCA on a FacileDataStore. If you want to run PCA on a
subset of samples, pass in a facile_frame sample descriptor
}
\examples{
if (interactive()) {
efds <- exampleFacileDataSet()
# run tumor vs normal comparisons vs each, then run compare9) on the results
pca.crc <- efds \%>\%
  filter_samples(indication == "CRC") \%>\%
  fpcaGadget()
report(pca.crc)
shine(pca.crc)
}
}
