% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R, R/fpca.R
\name{fplot}
\alias{fplot}
\alias{fplot.FacilePCA}
\title{Creates an interactive plot of the immersive result}
\usage{
fplot(x, ...)

\method{fplot}{FacilePCA}(x, pcs = 1:3, title = "Facile PCA",
  hover = NULL, color_aes = NULL, color_map = NULL, shape_aes = NULL,
  shape_map = NULL, size_aes = NULL, size_map = NULL, ...)
}
\description{
Creates an interactive plot of the immersive result
}
