% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freduce_dims.R
\name{freduce_dims}
\alias{freduce_dims}
\title{General dimensionality reduction}
\usage{
freduce_dims(x, method = .reduce_dim_methods(), ...)
}
\value{
a \code{FacileReducedDimsResult}
}
\description{
General dimensionality reduction
}
