% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frunGadget.R
\name{frunGadget}
\alias{frunGadget}
\title{Takes care of the tedium of wrapping a facilemodule to run as a gadget.}
\usage{
frunGadget(analysisModule, analysisUI, x, user = Sys.getenv("USER"),
  title = "Facile Analysis Gadget", height = 600, width = 1000,
  viewer = "dialog", ..., debug = FALSE)
}
\description{
We will implement barebones analysis gadgets that delegate down into here.
}
