% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsample_similarity.R
\name{fsample_similarity}
\alias{fsample_similarity}
\title{Run correlation analysis across a set of samples.}
\usage{
fsample_similarity(
  x,
  assay_name = NULL,
  cor_method = "pearson",
  dist_method = "euclidean",
  use = "everything",
  features = NULL,
  filter = "variance",
  ntop = 1000,
  col_covariates = NULL,
  batch = NULL,
  main = NULL,
  ...,
  metadata = list()
)
}
\arguments{
\item{x}{a facile_frame of samples}

\item{features}{what features to use?}

\item{filter}{default to "variance"}

\item{ntop}{default to top 1000}

\item{method}{"pearson" (default), or "spearman"}

\item{covariate}{if a string, then pull off the column from \code{x}}
}
\description{
This is a typical QC step for RNA-seq data, for instance, where you want
to show how well samples are correlated with each other, hopefully to see
that the samples from withint the same experimental group are better
correlated with each other than samples from other gruops
}
\examples{
afds <- FacileData::an_fds()
xs.all <- FacileData::samples(afds) |> FacileData::with_sample_covariates()
xs <- xs.all |> dplyr::filter(cond == "DKD")
ss <- fsample_similarity(xs)
viz(ss, "corr", annotate = "cell_abbrev")
viz(ss, "dist", annotate = "cell_abbrev")
}
