% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fscatterplot.R
\name{fscatterplot}
\alias{fscatterplot}
\alias{fscatterplot.data.frame}
\alias{fscatterplot.tbl}
\title{Draws interactive 2 or 3d scatterplots over data}
\usage{
fscatterplot(x, axes, color_aes = NULL, color_map = NULL,
  shape_aes = NULL, shape_map = NULL, size_aes = NULL, size_map = NULL,
  hover_aes = NULL, hover_map = NULL, hover = NULL, ...)

\method{fscatterplot}{data.frame}(x, axes, color_aes = NULL,
  color_map = NULL, shape_aes = NULL, shape_map = NULL, size_aes = NULL,
  size_map = NULL, hover_aes = NULL, hover_map = NULL, face_aes = NULL,
  hover = NULL, ...)

\method{fscatterplot}{tbl}(x, axes, color_aes = NULL, color_map = NULL,
  shape_aes = NULL, shape_map = NULL, size_aes = NULL, size_map = NULL,
  hover_aes = NULL, hover_map = NULL, face_aes = NULL, hover = NULL,
  ...)
}
\arguments{
\item{x}{a data object}

\item{axes}{the definition of the x,y,z axes}
}
\value{
a plotly object
}
\description{
Draws interactive 2 or 3d scatterplots over data
}
