% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsea.R
\name{fsea}
\alias{fsea}
\title{Performs Gene Set Enrichment Analyses}
\usage{
fsea(x, gdb, methods, ...)
}
\arguments{
\item{x}{A \code{FacileAnalysisResult} object}

\item{gdb}{A \code{multiGSEA::GeneSetDb} object}

\item{methods}{the GSEA methods to use on \code{x}.}
}
\value{
A FacileGSEAResult object, which includes a MultiGSEAResult object
as it's \code{result()}.
}
\description{
\strong{For now} only GSEA methods process a pre-ranked feature set work, like
\code{"cameraPR"} or \code{"fsea"}.
}
\section{Updates required to multiGSEA}{

I need to update multiGSEA to take an input data.frame of differential
expression statistics to work with goseq as well such that it doesn't have to
run the differential expression stuff again.

Once this is implemented, the a call to fsea will also work on an Anova
result, as well.
}

\examples{
gdb <- multiGSEA::getMSigGeneSetDb("h", "human", id.type = "entrez")

# GSEA from t-test result ---------------------------------------------------
ttest.res <- FacileData::exampleFacileDataSet() \%>\%
  FacileData::filter_samples(indication == "CRC") \%>\%
  fdge_model_def(covariate = "sample_type",
                 numer = "tumor", denom = "normal", fixed = "sex") \%>\%
  fdge(method = "voom")
ttest.gsea <- fsea(ttest.res, gdb)

# GSEA from ANOVA result ----------------------------------------------------
# Not yet implemented, requires small update to multiGSEA

# GSEA over loadings on a Principal Component -------------------------------
# Not yet implemented, need to get a signed weight out of eigenWeightedMean
# right now we just have weights. Or, fully extracing the biplot code, I
# think the weight should be the length on the PC of choice, and the sign is
# the same.
}
