% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdge-analysis-utils.R
\name{fvolcano}
\alias{fvolcano}
\title{Stylized volcano plot(s)}
\usage{
fvolcano(
  x,
  lfc = 1,
  fdr = 0.1,
  highlight = NULL,
  label = !is.null(highlight),
  label_max_overlaps = 20,
  with_stats = FALSE,
  nudge_x = 0,
  nudge_y = 0,
  interactive = FALSE,
  interactive_opts = list(),
  color.insig = "lightgrey",
  size.insig = 1,
  alpha.insig = 0.7,
  color.sig = "cornflowerblue",
  size.sig = 1,
  alpha.sig = 0.7,
  color.highlight = "firebrick2",
  size.highlight = 2.5,
  color.label = color.highlight,
  size.label = size.highlight,
  size.label_text = 4,
  xlims = "square",
  ylims = NULL,
  vline = NULL,
  vline_color = "firebrick",
  vline_type = "dashed",
  event_source = "A",
  ...,
  facet_ncol = 5,
  title = NULL,
  yaxis = c("padj", "pval"),
  contrast = NULL,
  dge_method = "DESeq2",
  grouping_var = "portfolio_id"
)
}
\arguments{
\item{x}{a \code{FacileTtestAnalysisResult} or a list of them to facet}

\item{lfc, fdrdefine}{significant result thresholds}

\item{highlight}{name or feature_id of genes to highlight. If the vector is
named, then the names are the genes to highlight and the values are the
colors, otherwise the values are the elements to highlight, and they will
be highlighted using the color specified by \code{color.highlight}}

\item{label}{If \code{TRUE} (default), then we add labels to the highlighted
genes using ggrepel}

\item{interactive}{If \code{TRUE} (default), the signficant and the highlighted
genes will be available for "interactivity" using ggiraph. If the
plot is plotted "normally", it is a standard ggplot, however you can call
\code{ggiraph::girafe(ggobj = this)}, and you'll get an interactive plot}

\item{color.insig, size.insig, alpha.insig}{aesthetics of points of insig genes}

\item{color.sig, size.sig, alpha.sig}{aesthetics of points of significant genes}

\item{color.highlight, size.highlight}{aesthetics of genes to highlight}

\item{xlims}{if \code{"square"} (default), will set x axis to be max(abs(logFC))}

\item{...}{stuff}
}
\description{
Stylized volcano plot(s)
}
\examples{
xs.all <- FacileData::an_fds() |>
  FacileData::samples() |> 
  FacileData::with_sample_covariates()
xs <- xs.all |> dplyr::filter(cond == "AKI")
dge <- flm_def(xs, "cell_abbrev", "IMM", "EC") |> fdge(method = "voom")

res <- tidy(dge) |> dplyr::arrange(pval)
hi <- res |> arrange(desc(logFC)) |> head(100) |> sample_n(20)
lbl <- head(res, 5)

fvolcano(
  dge,
  lfc = log2(2.5),
  fdr = 0.1,
  highlight = "ENSG00000067113"
)

fvolcano(
  dge,
  lfc = log2(2.5),
  fdr = 0.05,
  highlight = hi,
  label = lbl
)

fvolcano(
  dge,
  lfc = log2(2.5),
  fdr = 0.1,
  highlight = c(PLPP1 = "red", ROBO4 = "pink", LCP1 = "orange")
)


fvolcano(
  dge,
  lfc = log2(2.5),
  fdr = 0.1,
  highlight = head(res, 5)
)

fvolcano(
  dge,
  lfc = log2(2.5),
  fdr = 0.1,
  highlight = head(res, 5),
  label = FALSE
)

fvolcano(
  dge,
  lfc = log2(2.5),
  fdr = 0.1,
  highlight = head(res, 5),
  label = TRUE
)

fvolcano(
  dge,
  lfc = log2(2.5),
  fdr = 0.1,
  highlight = head(res, 5),
  label = "HCST"
)
}
