% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffsea-base.R
\name{geneset}
\alias{geneset}
\title{Extract geneset from ffsea result}
\usage{
geneset(x, name, collection = NULL, ...)
}
\arguments{
\item{x}{the \code{FacileFseaAnalysisResult}}

\item{name}{the name of the geneset (string)}

\item{collection}{(optional) often the name of the geneset is unique. If not
then the collection,name must be unique by GeneSetDb anyway.}

\item{...}{passed down to sparrow::geneSet}
}
\description{
This passes through sparrow functionality, but those are S4 methods, and we
living in an s3 world here ...
}
