% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffsea-base.R
\name{genesets}
\alias{genesets}
\title{Extract table of genesets used in analysis}
\usage{
genesets(x, name, collection = NULL, ...)
}
\arguments{
\item{x}{the \code{FacileFseaAnalysisResult}}

\item{...}{passed down to sparrow::geneSets}
}
\description{
This passes through sparrow functionality, but those are S4 methods, and we
living in an s3 world here ...
}
