% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frunGadget.R
\name{interacted}
\alias{interacted}
\title{Extracts interactions over an object that happened in a gadget}
\usage{
interacted(x, ...)
}
\arguments{
\item{x}{any object}
}
\value{
a named list of interactions recorded from running \code{\link[=frunGadget]{frunGadget()}}
}
\description{
These are stored and returned as a named list
}
