% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffsea-shiny.R
\name{ffseaGadget}
\alias{ffseaGadget}
\alias{ffseaRun}
\title{Full interactive executation of GSEA, soup to nuts.}
\usage{
ffseaGadget(x, title = "Feature Set Enrichment Analysis", height = 800,
  width = 1000, ...)

ffseaRun(input, output, session, rfds, aresult, gdb = NULL, ...,
  debug = FALSE)
}
\arguments{
\item{x}{A FacileAnalysisResult that has an implemented \code{ffsea.*} method}

\item{aresult}{A \code{FacileAnalysisResult} that has a \code{ffsea.*} method defined.}

\item{gdb}{A GeneSetDb object to use for testing.}
}
\description{
For now, feature set enrichment analysis are only performed downstream of
a \code{FacileAnalysisResult}. This requires that we have an ffsea method defined
for the specific class of the result, ie. \code{ffsea.FacilePcaAnalysisResult}.
}
\section{Custom Run Configuration}{

The options presented to the user for running a feature set enrichment
analysis on a \code{FacilePcaAnalysisResult} will be different than the ones
made available for an enrichment analysis over a \code{FacileTtestAnalysisResult}
or even a \code{FacileAnovaAnalysisResult}.

As such, each type of result should define a UI that accepts the appropriate
parameters for its corresponding \code{ffsea.*} method, and a server function
that extract and invokes the function.
}

\examples{
gdb <- multiGSEA::getMSigGeneSetDb("h", "human", id.type = "entrez")
dge.ttest <- FacileData::exampleFacileDataSet() \%>\%
  FacileData::filter_samples(indication == "CRC") \%>\%
  fdge_model_def(covariate = "sample_type",
                 numer = "tumor", denom = "normal", fixed = "sex") \%>\%
  fdge(method = "voom")
}
