% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipca.R
\name{ipca}
\alias{ipca}
\alias{ipca.DGEList}
\alias{ipca.matrix}
\alias{iplot.ipca}
\title{Immersive PCA}
\usage{
ipca(x, pcs = 1:10, ntop = 500, row_covariates = NULL,
  col_covariates = NULL, ...)

\method{ipca}{DGEList}(x, pcs = 1:10, ntop = 500,
  row_covariates = x$genes, col_covariates = x$samples,
  prior.count = TRUE, log = TRUE, ...)

\method{ipca}{matrix}(x, pcs = 1:10, ntop = 500, row_covariates = NULL,
  col_covariates = NULL, ...)

\method{iplot}{ipca}(x, pcs = 1:3, color_aes = NULL, shape_aes = NULL,
  hover = NULL, title = "Immersive PCA", ...)
}
\arguments{
\item{x}{a data container}
}
\value{
an ipca result
}
\description{
The code here is largely inspired by DESeq2's plotPCA.
}
\details{
You should look at factominer:
\itemize{
\item http://factominer.free.fr/factomethods/index.html
\item http://factominer.free.fr/graphs/factoshiny.html
}
}
