% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages-warnings-errors.R
\name{messages}
\alias{messages}
\alias{warnings}
\alias{errors}
\title{Extracts messages, warnings, and error messages from a FacileAnalysis Result}
\usage{
messages(x, ...)

warnings(...)

errors(x, ...)
}
\description{
Extracts messages, warnings, and error messages from a FacileAnalysis Result
}
