% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ranks}
\alias{ranks}
\title{Extracts feature or sample ranks from a FacileAnalysis}
\usage{
ranks(x, ...)
}
\arguments{
\item{x}{A FacileAnalysisResult}
}
\description{
Among other thigns, analyses can often provide rankings over features or
samples. For instance, a differential expression result may provide ranking
over the genes. This generic extract ranks of one type or another.
}
