% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactiveGeneSetDb-module.R
\name{reactiveGeneSetDb}
\alias{reactiveGeneSetDb}
\title{A ReactiveGeneSetDb for use in the shiiny world.}
\usage{
reactiveGeneSetDb(
  input,
  output,
  session,
  gdb,
  min.gs.size = 2L,
  max.gs.size = Inf,
  default_collections = NULL,
  ...
)
}
\arguments{
\item{gdb}{A static or reactive GeneSetDb object}

\item{min.gs.size, max.gs.size}{the default minimum and maximum geneset size
set in the UI when \code{gdb} is first loaded or changes (when reactive)}
}
\description{
This can be instantiated from a "static" or "reactive" GeneSetDb object.
It allows users to customize which genesets are active by:
}
\details{
\enumerate{
\item Filtering out entire collections; and
\item Filtering genesets based on min and max (gene) size.
}

This should be in mutliGSEA.shiny, but is developed here for now to make
it easier to develop within the FacieAnalysis shiny world.
}
