% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{redo}
\alias{redo}
\title{Redoes the analysis based on something new (samples, params, etc.)}
\usage{
redo(x, ..., samples = NULL)
}
\arguments{
\item{x}{The original FacileAnalysisResult}

\item{...}{probably additional arguments to retweak the original analysis}

\item{samples}{redo the analysis on a subset of these samples}
}
\description{
Redoes the analysis based on something new (samples, params, etc.)
}
