% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{samples}
\alias{features}
\alias{plot_data}
\alias{tidy}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{FacileData}{\code{\link[FacileData:features]{features()}}, \code{\link[FacileData:samples]{samples()}}}

  \item{FacileViz}{\code{\link[FacileViz:plot_data]{plot_data()}}}

  \item{broom}{\code{\link[broom:tidy]{tidy}}}
}}

