% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfds.R
\name{refds}
\alias{refds}
\title{Reconstitute FacileDataStore links from unfds'd objects}
\usage{
refds(x, fds, ...)
}
\arguments{
\item{fds}{a FacileDataStore}

\item{an}{object to reconstitute fds links to}
}
\description{
Not exported on purpose
}
