% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R, R/fpca.R
\name{render}
\alias{render}
\alias{render.FacilePCA}
\title{Creates an interactive plot of the FacileAnalysisResult}
\usage{
render(x, ...)

\method{render}{FacilePCA}(x, pcs = 1:3, color_aes = NULL,
  color_map = NULL, shape_aes = NULL, shape_map = NULL, size_aes = NULL,
  size_map = NULL, hover_aes = NULL, hover_map = NULL, hover = NULL,
  ...)
}
\description{
Creates an interactive plot of the FacileAnalysisResult
}
