% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{result}
\alias{result}
\title{Retrieves the main (or alternative) result from a FacileAnalysis}
\usage{
result(x, name = "result", ...)
}
\arguments{
\item{x}{A FacileAnalysisResult}

\item{name}{the name of the result to retrieve. Default is \code{"result"}, which
should represent the top-level result for all FacileAnalysisResult objects.}
}
\description{
This should always return a tidy data frame of one sort or another.
}
\details{
FacileAnalysisResults should should all aslo have a top-level result
accessible via obj[\link{"result"}]
}
