% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization.R
\name{fsave}
\alias{fsave}
\alias{fload}
\title{Save and load FacileAnalysisResult objects to/from disk.}
\usage{
fsave(x, file, with_fds = FALSE, ...)

fload(file, fds = NULL, with_fds = TRUE, ...)
}
\arguments{
\item{file}{A path to a file (should support \code{base::connections} in the
future.}

\item{with_fds}{Serialize the FacileDataStore with the object? Default is
\code{FALSE} and you should have a good reason to change this behavior.}

\item{fds}{The \code{FacileDataStore} the object was run on.}

\item{A}{\code{FacileAnalysisResult} object.}
}
\value{
\code{NULL} for \code{fsave}, the (correctly sublcassed) \code{FacileAnalysisResult}
for \code{fload}, along with \code{FacileDataStore} (\code{fds}) attached
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

Saving and loading FacileAnalysisResults can be tricky because they rely
on a chain of parent objects that created them, all of which have a reference
to the FacileDataStore that they came from.

Currenly, the FacileDataStore is removed from every internal object when
before it is serialized via \code{fsave()}. The FacileDataStore is then restored
to "just where it needs to be" (using strong assumptions of where it needs
to be, instead of logging where it was removed from in \code{fsave()}) when
brought back to life via \code{fload()}.
}
\details{
An \code{fsave_info} attribute will be attached to the serialized object before
saving which will be a list that holds information required to materialize
the analysis result successfully. Minimally this will include info about how
to reconstitute and attach the FacileDataStore that the object was generated
from.

Elements of \code{fsave_info} list so far include:
\itemize{
\item \code{fds_class}: A string indicating the class of the FacileDataStore that the
result \code{x} was generated from.
\item \code{fds_path}: The path on the filesystem that teh FacileDataStore can be
found. Currently we expect this to be a FacileDataSet
\item \code{fds_anno_dir}: The annotation directory for the FacileDataSet
\item \code{fds_add}: A descriptor that indicate which elements in the serialized
object should have the fds attached to them, besides itself.
}
}
