% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{signature}
\alias{signature}
\title{Converts an analysis result into a "signature" type of thing.}
\usage{
signature(x, ...)
}
\arguments{
\item{x}{A \code{FacileAnalysisResult}.}
}
\description{
This is mainly for feature rankings, but could be something else? Returns
a table of features that can be piped into GeneSetDb() constructor, should
insipiration strike.
}
\details{
Most of the \code{signature.*} functions defined over the result of analysis first
passes through a call to \code{ranks()} first, which then calls the \code{signature()}
method on those ranks, ie. the two segments below will produce the same
result.\preformatted{sigs1 <- FacileData::exampleFacileDataSet() \%>\%
  FacileData::filter_samples(indication == "CRC") \%>\%
  fpca() \%>\%
  signature()
sigs2 <- FacileData::exampleFacileDataSet() \%>\%
  FacileData::filter_samples(indication == "CRC") \%>\%
  fpca() \%>\%
  ranks() \%>\%
  signature()
}
}
\examples{
pca.sigs <- FacileData::exampleFacileDataSet() \%>\%
  FacileData::filter_samples(indication == "CRC") \%>\%
  fpca() \%>\%
  signature()
}
