% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{signed}
\alias{signed}
\title{Are these results (ranks, signatures) signed?}
\usage{
signed(x, ...)
}
\arguments{
\item{x}{a rank or signature result}
}
\value{
logical
}
\description{
Are these results (ranks, signatures) signed?
}
