% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{status}
\alias{status}
\title{Reports the status of the AnalysisResult}
\usage{
status(x, type = "message", ...)
}
\arguments{
\item{x}{A \code{FacileAnalysisResult}}
}
\value{
a FacileAnalysisResultStatus (message)
}
\description{
Reports the status of the AnalysisResult
}
