% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdge-analysis-utils.R
\name{summarize_fdge}
\alias{summarize_fdge}
\title{Summarize a DGE result (data.frame) with dge counts and rank positions.}
\usage{
summarize_fdge(
  x,
  fdr = 0.1,
  lfc = 1,
  rank = NULL,
  only_significant = TRUE,
  ...,
  contrast = NULL
)
}
\description{
For genes enumerated in \code{rank}, the user is given the option to
filter/include those based on their nominal pvalues, which is controlled
by the \code{rank_stats} option. By default, nominal pvalue is used for ranking.
}
\examples{
xs.all <- FacileData::an_fds() |>
  FacileData::samples() |> 
  FacileData::with_sample_covariates()
xs <- xs.all |> dplyr::filter(cond == "AKI")
dge <- flm_def(xs, "cell_abbrev", "IMM", "EC") |> fdge(method = "voom")
summarize_fdge(dge, lfc = 1, fdr = 0.05)
(s1 <- summarize_fdge(dge, rank = c("LCP1", "SOX18")))
}
