% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdge-analysis-utils.R
\name{summarize_fdge_all}
\alias{summarize_fdge_all}
\title{Tally the number of up/down regulated genes across a series of
comparisons.}
\usage{
summarize_fdge_all(
  comps,
  dge_res,
  fdr = 0.1,
  lfc = 0,
  rank = NULL,
  ...,
  dge_name = "dge",
  dge_label = "label"
)
}
\arguments{
\item{comps}{a data.frame of comparisons that were run, with a \verb{$label}
column}

\item{fdr}{the FDR (padj) threshold for significance}

\item{lfc}{the minimum thrshold for significance. If you are summarizing
a treat_lfc result, you will want to keep this at 0 because the threshold
is already part of the test.}

\item{dge.res}{a list of dge results, or list of lists with dge results
in them. I sometimes iterate over each comparisons and have for each one
a \code{dge} result, a \code{lfc} result (test against a threshold with treat),
a \code{gsea} result, etc. If \code{dge.name} is provided, we assume \code{dge.res} is
a list of lists and try to fish out the result with \code{dge.name}}

\item{rank.down, rank.up}{Report the position of the genes listed here in the
set of statistically significant down (or up) regulated genes. Use gene
symbols for now.}

\item{dge.label}{the name of the colum in \code{comps} that stores the "label"
of the result. We will assume \code{dge.res} is named by these values.}
}
\description{
Tally the number of up/down regulated genes across a series of
comparisons.
}
\examples{
xs.all <- FacileData::an_fds() |>
  FacileData::samples() |> 
  FacileData::with_sample_covariates()
xs <- xs.all |> dplyr::filter(cond == "AKI")
comps <- dplyr::tibble(
  covariate = "cell_abbrev",
  numer = c("IMM", "EC"),
  denom = c("CNT", "IMM"),
  label = sprintf("\%s_vs_\%s", numer, denom)
)
fdge.all <- lapply(1:nrow(comps), function(i) {
  info <- comps[i,]
  xs |> 
    flm_def(info$covariate, numer = info$numer, denom = info$denom) |> 
    fdge(method = "voom", metadata = as.list(info))
})
names(fdge.all) <- comps$label

fdsa <- summarize_fdge_all(comps, fdge.all, fdr = 0.1, lfc = log2(2))
rdsa <- summarize_dge_all(comps, rdge.all, fdr = 0.05, rank = c("HTT", "PMS1", "MSH3"))
}
\seealso{
\code{\link[=plot_dge_summary]{plot_dge_summary()}}
}
