% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{tidy.iresult}
\alias{tidy.iresult}
\title{Extract a tidy data.frame from an \code{iresult}}
\usage{
\method{tidy}{iresult}(x, ...)
}
\arguments{
\item{x}{an immersive result object (\code{iresult})}
}
\value{
a tidy data.frame of the results from an immersive analysis
}
\description{
Extract a tidy data.frame from an \code{iresult}
}
