% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfds.R
\name{unfds}
\alias{unfds}
\title{Strip and restore FacileDataStore pointers from/to FacileAnalysis Results}
\usage{
unfds(x, ...)
}
\arguments{
\item{x}{An object to remove the FacileDataStore attribute from}
}
\value{
a stripped out object
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

These functions are here to support serialization. \code{unfds} removes the link
to a FacileDataStore from any object, and \code{refds} restores it to only the
"blessed" objects.

These functions shouldn't really be used by users, but I'm having trouble
testing a non-exported S3 function in unit tests ...
}
