% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unreact.R
\name{unreact}
\alias{unreact}
\title{Replace ReactiveFacileDataStore versions with their "inert" version.}
\usage{
unreact(x, ...)
}
\arguments{
\item{x}{stuff and things}
}
\value{
a version of x with an "inert" FacileDataStore where the
ReactiveFacileDataStore once was
}
\description{
When analyses are run through a gadget, a ReactiveFacileDataStore is stored
everywhere a FacileDataStore would be (which is to say: in a lot of places).
This function tries to dig through the embedded results of an analysis and
replaces the ReactiveFaileDataStore objects with their inner FacileDataStore.
}
\details{
This function needs to be called within the gadget, ie. I'm pretty sure it
needs to fire in a reactive environment, which you get in the
\code{observeEvent(intput$done, { ... })} expression.
}
