% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userDefinedGeneSetDb-module.R
\name{userDefinedGeneSetDb}
\alias{userDefinedGeneSetDb}
\title{Creates a GeneSetDb from a user-specfied gene set definition}
\usage{
userDefinedGeneSetDb(input, output, session, ...)
}
\description{
This module provides an upload button that allows a user to upload a
table of gene set definitions in \code{\link[mutiGSEA:GeneSetDb]{mutiGSEA::GeneSetDb()}} format. Minimal
validation checks are implemented.
}
