% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics.R
\name{with_color}
\alias{with_color}
\alias{with_color.data.frame}
\alias{with_color.tbl}
\title{Augment a facile object with a color aesthetic}
\usage{
with_color(x, aesthetic, aes_map = NULL, ..., .default_color = I("black"))

\method{with_color}{data.frame}(x, aesthetic = NULL, aes_map = NULL,
  out_column = ".color_aes.", ..., .default_color = I("black"))

\method{with_color}{tbl}(x, aesthetic = NULL, aes_map = NULL,
  out_column = ".color_aes.", ..., .default_color = I("black"))
}
\arguments{
\item{x}{an object to colorize (currently just data.frame or tbls)}

\item{aesthetic}{an "attribute" from \code{x} to color. The types of object this
argument can be is specific to the class of \code{x}. For instance, for a
\code{data.frame}, \code{aesthetic} would be a character vector that specifices
the column(s) used to map observations to colors.}

\item{aes_map}{the "logic" to use to map \code{aesthetic} to color. This can be
something like the name of an RColorBrewer map.}

\item{...}{arguments to pass down to specific implementations}

\item{.default_color}{the default color used when only one level exists}

\item{out_column}{the name of the column that is appended to \code{x} that will
hold the colors}

\item{...}{moar things}
}
\value{
the object \code{x} with an upated \code{aes_map(x, "color")}

\code{x} with an additional \code{out_column} column that holds the colors for
the observations in the rows of \code{x}
}
\description{
This function augments a variety of object with color mappings. For now we
assume objects are only data.frames or tbls.
}
\section{data.frames and tibbles}{


data.frames and tibbles are augmented with:
\enumerate{
\item A \code{.color_aes.variable} column that hold the value of the element "level"
for the color mapping
\item \code{.color_aes.value} column that holds the actual color for the observation
\item an updated aes_map(\code{x}, "color")` color map that has the color map
}

The reason we add a \code{.color_aes.variable} is because the user may decide
to color observations by combinations of columns, not just a single column.
}

\examples{
x <- tibble(
  category = sample(c("a", "b", "c"), 20, replace = TRUE),
  subcat = sample(c("y", "z"), 20, replace = TRUE),
  score = rnorm(20))
x <- with_color(x, aesthetic = "category", "Set3")
y <- with_color(x, aesthetic = "category", "Set1")
}
\seealso{
\code{\link[=aes_map]{aes_map()}}
}
