% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{mutate.FacileAnalysisResult}
\alias{mutate.FacileAnalysisResult}
\alias{select.FacileAnalysisResult}
\alias{filter.FacileAnalysisResult}
\alias{summarise.FacileAnalysisResult}
\alias{summarize.FacileAnalysisResult}
\alias{arrange.FacileAnalysisResult}
\alias{group_by.FacileAnalysisResult}
\title{ImmersiveAnalysis result objects can be maniuplated with dplyr verbs}
\usage{
\method{mutate}{FacileAnalysisResult}(.data, ...)

\method{select}{FacileAnalysisResult}(.data, ...)

\method{filter}{FacileAnalysisResult}(.data, ...)

\method{summarise}{FacileAnalysisResult}(.data, ...)

\method{summarize}{FacileAnalysisResult}(.data, ...)

\method{arrange}{FacileAnalysisResult}(.data, ...)

\method{group_by}{FacileAnalysisResult}(.data, ...)
}
\description{
ImmersiveAnalysis result objects can be maniuplated with dplyr verbs
}
