% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{mutate.FacileAnalysis}
\alias{mutate.FacileAnalysis}
\alias{select.FacileAnalysis}
\alias{filter.FacileAnalysis}
\alias{summarise.FacileAnalysis}
\alias{summarize.FacileAnalysis}
\alias{arrange.FacileAnalysis}
\alias{group_by.FacileAnalysis}
\title{ImmersiveAnalysis result objects can be maniuplated with dplyr verbs}
\usage{
\method{mutate}{FacileAnalysis}(.data, ...)

\method{select}{FacileAnalysis}(.data, ...)

\method{filter}{FacileAnalysis}(.data, ...)

\method{summarise}{FacileAnalysis}(.data, ...)

\method{summarize}{FacileAnalysis}(.data, ...)

\method{arrange}{FacileAnalysis}(.data, ...)

\method{group_by}{FacileAnalysis}(.data, ...)
}
\description{
ImmersiveAnalysis result objects can be maniuplated with dplyr verbs
}
