% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdge.R
\name{fdge}
\alias{fdge}
\title{Peforms a differential expression analysis and GSEA.}
\usage{
fdge(x, design = NULL, contrast = NULL, covariates = NULL,
  filter = NULL, method = .dge_methods, gsea = "camera",
  gdb = NULL, ...)
}
\arguments{
\item{x}{a data source}

\item{covariates}{a data.frame of covariates (columns) for each
sample (rows) in \code{x}}

\item{numer}{character vector defining the covariate/groups that
make up the numerator}

\item{denom}{character vector defining the covariate/groups that
make up the denominator}

\item{fixed}{character vector defining the covariate/groups to
use as fixed effects}
}
\description{
Peforms a differential expression analysis and GSEA.
}
