% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdge-interact.R, R/fdge.R
\name{report.FacileTtestDGEResult}
\alias{report.FacileTtestDGEResult}
\alias{fdge}
\alias{fdge.FacileAnovaModelDefinition}
\alias{fdge.FacileDGEModelDefinition}
\title{Peforms a differential expression analysis and GSEA.}
\usage{
\method{report}{FacileTtestDGEResult}(x, type = c("dge", "features"),
  ntop = 200, max_padj = 0.1, min_logFC = 1, features = NULL,
  round_digits = 3, event_source = "A", webgl = TRUE,
  caption = NULL, ...)

fdge(x, ...)

\method{fdge}{FacileAnovaModelDefinition}(x, assay_name = NULL,
  method = NULL, filter = "default", ...)

\method{fdge}{FacileDGEModelDefinition}(x, assay_name = NULL,
  method = NULL, filter = "default", treat_lfc = NULL, ...)
}
\arguments{
\item{x}{a data source}

\item{numer}{character vector defining the covariate/groups that
make up the numerator}

\item{denom}{character vector defining the covariate/groups that
make up the denominator}

\item{fixed}{character vector defining the covariate/groups to
use as fixed effects}

\item{covariates}{a data.frame of covariates (columns) for each
sample (rows) in \code{x}}
}
\description{
Peforms a differential expression analysis and GSEA.
}
\section{Interacting with results}{


The \code{report} function will create an htmlwidget which can be explored by
the analyst or dropped into an Rmarkdown report.

\code{report(result, "dge", max_padj = 0.05, min_logFC = 1)} will create a
side-by-side volcano and datatable for differential expression results.
}

\examples{
efds <- FacileData::exampleFacileDataSet()
samples <- FacileData::filter_samples(efds, indication == "BLCA")
mdef <- fdge_model_def(samples, covariate = "sample_type",
                       numer = "tumor", denom = "normal")
dge <- fdge(mdef, method = "voom")

if (interactive()) {
  viz(dge)
  shine(dge)
}
}
