% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdge-shiny.R
\name{fdgeAnalysis}
\alias{fdgeAnalysis}
\title{Wrapper module to perform and interact with differential a expression result}
\usage{
fdgeAnalysis(input, output, session, rfds, ...)
}
\arguments{
\item{model_def}{the module returned from \code{\link[=fdgeModelDefModule]{fdgeModelDefModule()}}}
}
\value{
a \code{ShinyFacileDGEResult}, the output from \code{\link[=fdge]{fdge()}}
}
\description{
Wrapper module to perform and interact with differential a expression result
}
