% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdge-ffsea-analysis-module.R
\name{fdgeseaGadget}
\alias{fdgeseaGadget}
\title{Gadget to run both DGE and GSEA}
\usage{
fdgeseaGadget(x, title = "DGE and GSEA", height = 800, width = 1000,
  ...)
}
\description{
Gadget to run both DGE and GSEA
}
