% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea.R
\name{fgsea}
\alias{fgsea}
\title{Performs Gene Set Enrichment Analyses}
\usage{
fgsea(x, ..)
}
\description{
For now we only support GSEA methods that can be run from a
\code{FacileDGETestResult}, such as \code{"cameraPR"} and \code{"goseq"}
}
