% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpca.R
\name{fpca}
\alias{fpca}
\alias{fpca.DGEList}
\alias{fpca.matrix}
\title{Runs a facile PCA}
\usage{
fpca(x, pcs = 1:10, ntop = 500, row_covariates = NULL,
  col_covariates = NULL, ...)

\method{fpca}{DGEList}(x, pcs = 1:10, ntop = 500,
  row_covariates = x$genes, col_covariates = x$samples,
  prior.count = 3, log = TRUE, ...)

\method{fpca}{matrix}(x, pcs = 1:10, ntop = 500,
  row_covariates = NULL, col_covariates = NULL, ...)
}
\arguments{
\item{x}{a data container}
}
\value{
an fpca result
}
\description{
The code here is largely inspired by DESeq2's plotPCA.
}
\details{
You should look at factominer:
\itemize{
\item http://factominer.free.fr/factomethods/index.html
\item http://factominer.free.fr/graphs/factoshiny.html
}
}
