% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpca.R
\name{fpca}
\alias{fpca}
\alias{fpca.facile_frame}
\alias{fpca.matrix}
\title{Runs a facile PCA}
\usage{
fpca(x, pcs = 1:10, ntop = 500, row_covariates = NULL,
  col_covariates = NULL, ...)

\method{fpca}{facile_frame}(x, pcs = 1:10, ntop = 500,
  row_covariates = NULL, col_covariates = NULL, assay_name = NULL,
  custom_key = Sys.getenv("USER"), ...)

\method{fpca}{matrix}(x, pcs = 1:10, ntop = 500,
  row_covariates = NULL, col_covariates = NULL, ...)
}
\arguments{
\item{x}{a data container}
}
\value{
an fpca result
}
\description{
The code here is largely inspired by DESeq2's plotPCA.
}
\details{
You should look at factominer:
\itemize{
\item http://factominer.free.fr/factomethods/index.html
\item http://factominer.free.fr/graphs/factoshiny.html
}
}
\section{FacileDataStore (facile_frame)}{

We enable the user to supply extra sample covariates that are not found
in the FacileDataStore associated with these samples \code{x} by adding them as
extra columns to \code{x}.

If manually provioded col_covariates have the same name as internal sample
covariates, then the manually provided ones will supersede the internals.
}

\examples{
# Generate a data matrix with a sister covariate table
mat <- local({
  m <- matrix(rnorm(100 * 10), nrow = 100)
  colnames(m) <- letters[1:10]
  rownames(m) <- head(unique(
    replicate(200, paste(sample(letters, 5), collapse = ""))),
    nrow(m))
  m
})

# generate some sample (column) covariates
pdat <- local({
  pd <- FacileAnalysis:::example_aes_data_table(10, n.cats = 5)
  pd <- as.data.frame(pd)
  rownames(pd) <- colnames(mat)
  pd
})

# analyze and vizualize
res <- fpca(mat, col_covariates = pdat)
if (interactive()) {
  plot(res, color_aes = "category")
  report(res, color_aes = "category", caption = "Check this out")
}
}
