% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{param}
\alias{param}
\title{Extract the value of a parameter(s) used in a FacileAnalysis result}
\usage{
param(x, name = NULL, ...)
}
\arguments{
\item{x}{A FacileAnalysisResult}

\item{name}{the name of the parameter to extract}
}
\description{
Extract the value of a parameter(s) used in a FacileAnalysis result
}
