% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tidy}
\alias{\%>\%}
\alias{callModule}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{broom}{\code{\link[broom]{tidy}}}

  \item{FacileData}{\code{\link[FacileData]{\%>\%}}}

  \item{shiny}{\code{\link[shiny]{callModule}}}
}}

