% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{samples}
\alias{features}
\alias{plot_data}
\alias{tidy}
\alias{\%>\%}
\alias{callModule}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{broom}{\code{\link[broom:reexports]{tidy}}}

  \item{FacileData}{\code{\link[FacileData:reexports]{\%>\%}}, \code{\link[FacileData]{features}}, \code{\link[FacileData]{samples}}}

  \item{FacileViz}{\code{\link[FacileViz]{plot_data}}}

  \item{shiny}{\code{\link[shiny]{callModule}}}
}}

