% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization.R
\name{fsave}
\alias{fsave}
\alias{fload}
\title{Save and load FacileAnalysisResult objects to/from disk.}
\usage{
fsave(x, file, with_fds = FALSE, ...)

fload(x, fds = NULL, anno = NULL, with_fds = TRUE, ...)
}
\arguments{
\item{x}{A \code{FacileAnalysisResult} object.}

\item{file}{A directory named \code{file} will be created to serialize the bits
required to reload this analysis result in another session. If \code{file}
already exists, this function will throw an error.}

\item{with_fds}{Serialize the FacileDataStore with the object? Default is
\code{FALSE} and you should have a good reason to change this behavior.}

\item{fds}{The \code{FacileDataStore} the object was run on.}
}
\value{
\code{NULL} for \code{fsave}, the (correctly sublcassed) \code{FacileAnalysisResult}
for \code{fload}, along with \code{FacileDataStore} (\code{fds}) attached
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Saving and loading FacileAnalysisResults can be tricky because they rely
on a chain of parent objects that created them, all of which have a reference
to the FacileDataStore that they came from.

Currenly, the FacileDataStore is removed from every internal object when
before it is serialized via \code{fsave()}. The FacileDataStore is then restored
to "just where it needs to be" (using strong assumptions of where it needs
to be, instead of logging where it was removed from in \code{fsave()}) when
brought back to life via \code{fload()}.
}
\details{
fsave creates a directory with two elements, the actually analysis result
(a qs2 object), and a \code{meta.yaml} that has metadata related to the
FacileAnalysisResult object being saved. The details in the \code{meta.yaml}
file will include the \code{metadata()} from \code{x}, and details about what
FacileDataStore was used to serialize the result.

metadata about the analysis result will be stored under the \verb{result:}
section, and the \verb{datastore:} section will hold the elemnts that were
previously associated with the \code{fsave_info} attribute from the v1
*.qs object

OLD:
An \code{fsave_info} attribute will be attached to the serialized object before
saving which will be a list that holds information required to materialize
the analysis result successfully. Minimally this will include info about how
to reconstitute and attach the FacileDataStore that the object was generated
from.

Elements of \code{fsave_info} list so far include:
\itemize{
\item \code{fds_class}: A string indicating the class of the FacileDataStore that the
result \code{x} was generated from.
\item \code{fds_path}: The path on the filesystem that teh FacileDataStore can be
found. Currently we expect this to be a FacileDataSet
\item \code{fds_anno_dir}: The annotation directory for the FacileDataSet
\item \code{fds_add}: A descriptor that indicate which elements in the serialized
object should have the fds attached to them, besides itself.
}
}
