% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unreact.R
\name{unreact}
\alias{unreact}
\title{Replace reactive versions of facile objects with their "inert" counterparts}
\usage{
unreact(x, ...)
}
\arguments{
\item{x}{stuff and things}
}
\value{
a version of x with an "inert" FacileDataStore where the
ReactiveFacileDataStore once was
}
\description{
This function is primariy meant to be called within the \code{frunGadget}
framework. We attempt to find all "reactive"-like components and covert them
into their "inert" versions so that they can be used in the context of an
ongoing analysis.
}
\details{
We are primariy concerned about replacing the \code{ReactiveFacileDataStore} that
is stashed within facile components (\code{facile_frames}) (set and retrieved with
the \code{set_fds()} and \code{fds()} setters and getters), with their inner
\code{FacileDataStore}

This function needs to be called within the gadget, ie. I'm pretty sure it
needs to fire in a reactive environment, which you get in the
\code{observeEvent(intput$done, { ... })} expression.
}
